feat: Allow multiple images for medications
This commit introduces the ability to add multiple images to a medication. It includes changes to the database, domain entities, and UI to support this feature. A new cropper page has been added to crop images before adding them. The BlocProvider has also been moved to main.dart for better organization.